requirements.txt
asn1crypto==0.24.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
dialogflow==0.5.2
feedparser==5.2.1
Flask==1.0.2
future==0.17.1
fuzzywuzzy==0.17.0
gnewsclient==1.11
google-api-core==1.9.0
google-auth==1.6.3
googleapis-common-protos==1.5.9
grpcio==1.20.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2